clean up utils files
