+++ bifrost/lib/turbolinks/util.ts
@@ -109,25 +109,52 @@ export function uuid() {
     .join("");
 }
 
-
-export function createScriptElement(element: Element): HTMLScriptElement {
+export function createScriptElement(element: Element, cb?: () => void) {
   if (element.getAttribute("data-turbolinks-eval") == "false") {
-    return element as HTMLScriptElement;
+    return element;
   } else {
     const createdScriptElement = document.createElement("script");
     createdScriptElement.textContent = element.textContent;
+    // async false makes scripts run in-order. it wont block js execution (thankfully)
+    //   https://github.com/turbolinks/turbolinks/issues/282#issuecomment-355731712
     createdScriptElement.async = false;
     copyElementAttributes(createdScriptElement, element);
+    if (cb) {
+      createdScriptElement.addEventListener("load", cb);
+    }
     return createdScriptElement;
   }
 }
 
-function copyElementAttributes(
+export function copyElementAttributes(
   destinationElement: Element,
   sourceElement: Element
 ) {
   for (const { name, value } of array(sourceElement.attributes)) {
     destinationElement.setAttribute(name, value);
+  }
+}
+
+export function getElementAttributes(element: Element) {
+  const bodyAttrs: Record<string, string> = {};
+  element.getAttributeNames().forEach((name) => {
+    bodyAttrs[name] = element.getAttribute(name)!;
+  });
+  return bodyAttrs;
+}
+
+function replaceElementWithElement(fromElement: Element, toElement: Element) {
+  const parentElement = fromElement.parentElement;
+  if (parentElement) {
+    return parentElement.replaceChild(toElement, fromElement);
+  }
+}
+export function activateNewBodyScriptElements(
+  newScriptElements: HTMLScriptElement[]
+) {
+  for (const inertScriptElement of newScriptElements) {
+    const activatedScriptElement = createScriptElement(inertScriptElement);
+    replaceElementWithElement(inertScriptElement, activatedScriptElement);
   }
 }
